requirements.txt
# Django jsonfield for output storage
django-jsonfield
# Django-tojson
-e git://github.com/davidek/django-tojson.git#egg=tojson
